a11y: Use desktop file name as app's accessible ID
As the QGuiApplication::desktopFileName doc [1] says:

> This is the file name, without the full path or the
> trailing ".desktop" extension of the desktop entry that
> represents this application according to the freedesktop
> desktop entry specification.

Since the desktop entry represents the application, also
use the desktop file name for the accessible ID of the
application object.

This e.g. makes Okular's application object report an
accessible ID of "org.kde.okular" instead of just "QApplication"
as the AT-SPI2 AccessibleId property on Linux.

[1] https://doc.qt.io/qt-6/qguiapplication.html#desktopFileName-prop

Pick-to: 6.8
Change-Id: I554da7baaf817866db9044128f73a9ea634cddbf
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>