Fix argument in call to updateSectionBeforeAfter
This argument is in the wrong format (visual instead of logical)
Likely it doesn't matter much when moveSection updates
the actual section. This however does more than just
the section (though I couldn't reproduce an error).
If we in the future allows an override of moveSection one way or
another this bug is likely to become more annoying, though.

Change-Id: Ic5eafeacbac318d75e45fb11ccc1d3027ecc45e9
Reviewed-by: Christian Ehrlicher <ch.ehrlicher@gmx.de>