Make QDirListing::const_iterator post-increment return void
std::weakly_incrementable doesn't require a std::input_iterator's
post-increment operator to return anything¹, so don't.

¹ https://eel.is/c++draft/iterator.concepts#iterator.concept.winc-14

The return value of post-increment on objects that model
input_iterator is partially-formed (as a copy of an iterator since
advanced), so any attempts to use it, apart from assigning a new
iterator or destroying it, are undefined, anyway.

Found in API-review.

Pick-to: 6.8
Change-Id: If0c07250cec4e48c6e5176cd23864651e8ec3d27
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>