tst_QPluginLoader: use qCpuHasFeature instead of __builtin_cpu_supports
Fixes a build problem on VxWorks because their runtime library isn't
libgcc.a and is thus missing the symbols that the Clang emits into the
test code when the intrinsic is used. This also better aligns the CPU
check with what QtCore does.

Drive-by add a comment explaining what this is and why Darwin is
excluded.

Drive-by clean up of the CMakeLists.txt, especially the no-ops.

Task-number: QTBUG-115777
Pick-to: 6.8
Change-Id: Ib15e33ce157f7e6d99d0530076de6a1024733ed6
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>