QtPrivate: add "defined in" comment to findByteArray/lastIndexOf(~~~, char)
It's customary to add such a comment when an inline function is not
defined in the same header it is declared.

As a drive-by, remove the pure attribute from lastIndexOf().
Q_DECL_PURE/CONST_FUNCTIONS are only useful for out-of-line functions.
For inline functions, the compiler sees the body and can determine
that they're pure or const (in the __attribute__((const)) sense)
itself.

Found in API-review.

Pick-to: 6.8
Change-Id: I27aa5cfaf21b1015a6c8b9e4773a1c2c5134a226
Reviewed-by: Juha Vuolle <juha.vuolle@qt.io>