Restore support for iOS 16 for Qt 6.8
Our general strategy is to support the latest two iOS releases,
as most people upgrade their iOS devices quite aggressively.

But given that iOS 18 is not out yet, and that people still
spend some time upgrading once it's out, we don't want to be
in a situation where we effectively only support iOS 17.

Fixes: QTBUG-128678
Pick-to: 6.8
Change-Id: I959794944807e5b6b5e56f22d3c57977bc4f2aee
Reviewed-by: Doris Verria <doris.verria@qt.io>