Don't mark Q{Explicitly.}SharedDataPointer as Qt_6_PRIVATE_API
The QT_DECLARE_Q{,E}SDP_SPECIALIZATION_DTOR_WITH_EXPORT macros make the
destrcutor of the particular instantiations be out-of-line and exported
from the library in question. In many cases, the class name that is
being instantiated can found in a _p.h and this causes syncqt to mark
any use of this class as private API, including this constructor that is
meant to be used from inline (user) code.

This patch marks all such instantiations to be "public" API. That will
include some uses that shouldn't be so marked, but this is easier than
having to do it piecemeal.

The ELF version was changed for thsee Symbols changed in QtGui (for example):
 QExplicitlySharedDataPointer<QPlatformPixmap>::~QExplicitlySharedDataPointer()
 QExplicitlySharedDataPointer<QColorSpacePrivate>::~QExplicitlySharedDataPointer()
 QExplicitlySharedDataPointer<QPageRangesPrivate>::~QExplicitlySharedDataPointer()
 QExplicitlySharedDataPointer<QColorTransformPrivate>::~QExplicitlySharedDataPointer()
 QExplicitlySharedDataPointer<QPdfOutputIntentPrivate>::~QExplicitlySharedDataPointer()@@Qt_6
 QExplicitlySharedDataPointer<QEventPointPrivate>::~QExplicitlySharedDataPointer()
 QExplicitlySharedDataPointer<QPageRangesPrivate>::~QExplicitlySharedDataPointer()
 QExplicitlySharedDataPointer<QColorTransformPrivate>::~QExplicitlySharedDataPointer()
 QExplicitlySharedDataPointer<QPenPrivate>::~QExplicitlySharedDataPointer()
 QExplicitlySharedDataPointer<QFontVariableAxisPrivate>::~QExplicitlySharedDataPointer()@@Qt_6
 QExplicitlySharedDataPointer<QPlatformPixmap>::~QExplicitlySharedDataPointer()
 QExplicitlySharedDataPointer<QColorSpacePrivate>::~QExplicitlySharedDataPointer()
 QExplicitlySharedDataPointer<QEventPointPrivate>::~QExplicitlySharedDataPointer()
 QExplicitlySharedDataPointer<QPenPrivate>::~QExplicitlySharedDataPointer()

Pick-to: 6.8
Task-number: QTBUG-128656
Task-number: QTBUG-117514
Change-Id: I1f25f1e61bb98159f9d3fffd63acbe04098cfef6
Reviewed-by: Dennis Oberst <dennis.oberst@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>