QIODevice: Add QIODevice::readLineInto()
The QByteArray QIODevice::readLine() method allocates new memory
(QByteArray) for each line.
Introduce 'QIODevice::readLineInto(QByteArray*, qint64 max)' which uses
pre-allocated memory to optimize performance.

Introduce QIODevicePrivate::skipLine(), similar to readLineData(), to
read a line without storing it.
Call it in readLineInto() when the line is nullptr and the maxsize is 0
or not specified.

[ChangeLog][QtCore][QIODevice] Added readLineInto() an alternative to
readLine() that utilizes pre-allocated memory.

Task-number: QTBUG-113547
Task-number: QTBUG-103108
Change-Id: I51f24b8664f80652bba3c8fc2cb3af4cf9adbeac
Reviewed-by: Marc Mutz <marc.mutz@qt.io>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>