QCollatorSortKey: mark compare() noexcept
The documentation says that sort keys can only be compared from the same
QCollator, so technically this is a non-wide contract. But there's no
way to even verify whether they were from the same QCollator or not in
three of the four implementations, so I'm discounting that case as a
narrowing and therefore declaring them all as wide-contract.

Change-Id: I7532825f0673e803ebb7fffd0e191e8acc10089e
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>