a11y: Add new BlockQuote role
Add a new QAccessible::BlockQuote role that can be used
to report quoted content to assistive technology
as such (e.g. by browsers or document editors like
LibreOffice).

This matches the role as specified in e.g. ARIA [1],
IAccessible2 [2] and AT-SPI [3].

Map the role to platform APIs as described in the
The W3C Core Accessibility mappings spec [4]:

* Linux/AT-SPI: ATSPI_ROLE_BLOCK_QUOTE
* Windows/UIA: UIA_GroupControlTypeId with a
  UIA_LocalizedControlTypePropertyId of
  "blockquote"
* macOS/NSAccessibility: NSAccessibilityGroupRole

In the QAccessible::Role enum, also add comments
for the other IAccessible2 roles that have been
added in the meantime, to match the existing approach.
(Note however that Qt itself doesn't even use
IAccessible2 for its Windows a11y bridge any more since
2017 commit 0cf6297c15be45d852be98c862bd0211e6de1aa2,
so it's a bit unclear to me whether there's really
value in keeping those IAccessible2 references there
in the long run.)

[1] https://w3c.github.io/aria/#blockquote
[2] https://github.com/LinuxA11y/IAccessible2/blob/2b8c2c79417bad4b464761a142fab45ffde8bfa8/api/AccessibleRole.idl#L318
[3] https://docs.gtk.org/atspi2/enum.Role.html#block_quote
[4] https://www.w3.org/TR/core-aam-1.2/#role-map-blockquote

[ChangeLog][QtGui][QAccessible] Added new BlockQuote
role that can be used to report quoted content as
such to assistive technology.

Fixes: QTBUG-128870
Change-Id: I8707d9d77640b67416f726d266b4cbec5ac09422
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>