Enable ctest junit XML test summary for the ASAN configuration
The ASAN configuration runs on OpenSUSE and has the variable

  CMAKE_MIN_SUPPORTED_BIN_PATH=/opt/cmake-3.16.8/bin

because in the non-ASAN configuration it prepends this to PATH to test
functionality with old CMake.

In the ASAN configuration the above variable is still set, but it's not
prepended to PATH, so the latest CMake is used. We now detect that
properly and enable junit XML summary files.

Change-Id: I75f7308b9332434853cbe2a9878f7b4c192710e7
Reviewed-by: Toni Saario <toni.saario@qt.io>