cmake: Allow skipping build tests via environment variable
On setups where the tests are built as part of the build phase, and then
copied over/mounted on a test nodes an run as a separate step, the test
node may not have the build tools installed.

In that setup we may want to skip the build tests, instead of seeing
them fail.

Ideally this would be possible as a combination of FIXTURES_REQUIRED
and a skipped FIXTURES_SETUP, but that combination does not result in
skipping the FIXTURES_REQUIRED tests:

 https://gitlab.kitware.com/cmake/cmake/-/issues/26125

Instead we ensure that all build tests are run via a wrapper script,
that emit a skip message if the test should be skipped. The CMake
tests then uses SKIP_REGULAR_EXPRESSION to skip the test. We prefer
this approach over SKIP_RETURN_CODE as the latter requires CMake
API, cmake_language(EXIT), that's only available in CMake 3.29.

For now the condition to skip the test is a hard-coded environment
variable, but in the future we could do detection of prerequisites.

Change-Id: Ie8ac4b7a74c264eac8bcc1e432c94510dd0f3a9c
Reviewed-by:  Alexey Edelev <alexey.edelev@qt.io>