QUrl/QDir: revert all recent changes to path normalization
This amends commit c023016ffdda9e2a27c2257ca85a8ffd384e8c08 to revert
all of the following commits:
214fbc658bc78ad39e0fb33daf2a3a3fa621b02b
492c646735faa2cad6d0459066dfd723b7ee99b0
804b0996746daae9cf3c901ce353261e1edd7728
e4576684681df8fe6d65a2f72379968c187db108
a1610c6c68ca2c9a3855a4e5e947033e6410d7c9

The reason for this is that they introduce a behavior change in QUrl,
causing the normalization/resolution of file:// directory paths not to
end in "/" and this behavior change revealed buggy code in QML that
concatenated strings expecting the slash to be there. Commit
c023016ffdda9e2 attempted to do that too, but it left QUrl in a state
that introduced a regression from 6.7.

Discussion on whether we will accept this behavior change (and require
fixes to buggy code) or keep the old behavior to avoid breakages will
happen for 6.8.x.

This commit reintroduces bug QTBUG-120396 to 6.8.0.

Task-number: QTBUG-120396
Fixes: QTBUG-128940
Change-Id: Id78899df1cea4321d9e1fffdcc0f406bbfd534b8
Reviewed-by: Jani Heikkinen <jani.heikkinen@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>